test: guard lazy file explorer sync policy
